// src/frontend/src/api/generated/types.gen.ts
export type ClimatiqDataVersionInformationRemoved = {
  status: "removed";
  replacedIn: string;
};

export type ClimatiqDataVersionInformationReplaced = {
  status: "replaced";
  replacedIn: string;
  replacedBy: string;
};

export type ClimatiqDataVersionInformationUpToDate = {
  status: "up_to_date";
};
// ...
export type ClimatiqEmissionFactor_Output = {
  source: "climatiq";
  attributes: Array<QualifiedAttribute>;
  data: ClimatiqEmissionFactorDetails;
  /**
   * The emission factor. If not passed, the climatiq API is accessed to retrieve the factor.
   */
  value?: {
    value: number;
    unit?: {
      /**
       * Unit symbol
       */
      symbol: string | null;
      /**
       * Long name of the unit
       */
      name: string | null;
      /**
       * The unit dimensionality, compatible with the `pint` python package
       */
      dim: {
        [key: string]: number;
      };
    } | null;
  } | null;
};

export type ClimatiqEmissionFactorDetails = {
  id: string;
  activityId: string;
  accessType: "private" | "public";
  name: string;
  category: string;
  sector: string;
  source: string;
  sourceLink: string;
  uncertainty?: number | null;
  year: number;
  yearReleased: number;
  region: string;
  regionName: string;
  description: string;
  unit: string;
  unitType: string;
  sourceLcaActivity: string;
  supportedCalculationMethods: Array<"ar4" | "ar5" | "ar6">;
  factor?: number | null;
  factorCalculationMethod?: "ar4" | "ar5" | "ar6" | null;
  constituentGases: {
    [key: string]: number | null;
  };
  dataVersionInformation:
    | ClimatiqDataVersionInformationUpToDate
    | ClimatiqDataVersionInformationReplaced
    | ClimatiqDataVersionInformationRemoved;
};
// ...
export type E2OEmissionRule_Output = {
  index: number;
  name?: string | null;
  activity: string;
  factor: LocalEmissionFactor_Output | ClimatiqEmissionFactor_Output;
  type: "E2OEmissionRule";
  objectType: string;
  qualifier?: string | null;
  readonly defaultName: string;
};

export type EmissionsState = {
  hasImportedEmissions?: boolean;
  hasRuleBasedEmissions?: boolean;
  hasObjectEmissions: boolean;
  readonly hasEmissions: boolean;
};

export type EventAttributeDefinition = {
  target: "event";
  name: string;
  unit?: {
    /**
     * Unit symbol
     */
    symbol: string | null;
    /**
     * Long name of the unit
     */
    name: string | null;
    /**
     * The unit dimensionality, compatible with the `pint` python package
     */
    dim: {
      [key: string]: number;
    };
  } | null;
  activity: string;
};

export type EventEmissionRule_Input = {
  index: number;
  name?: string | null;
  activity: string;
  factor: LocalEmissionFactor_Input | ClimatiqEmissionFactor_Input;
  /**
   * Disables validation of the result unit being a weight.
   */
  ignoreUnit?: boolean;
  type: "EventEmissionRule";
};

export type EventEmissionRule_Output = {
  index: number;
  name?: string | null;
  activity: string;
  factor: LocalEmissionFactor_Output | ClimatiqEmissionFactor_Output;
  type: "EventEmissionRule";
  readonly defaultName: string;
};
// ...
export type LocalEmissionFactor_Output = {
  source: "local";
  attributes: Array<QualifiedAttribute>;
  /**
   * A quantity (physical, currency, ...), containing a value and (optionally) a unit.
   */
  value: {
    value: number;
    unit?: {
      /**
       * Unit symbol
       */
      symbol: string | null;
      /**
       * Long name of the unit
       */
      name: string | null;
      /**
       * The unit dimensionality, compatible with the `pint` python package
       */
      dim: {
        [key: string]: number;
      };
    } | null;
  };
};
// ...
export type ObjectAttributeDefinition = {
  target: "object";
  name: string;
  unit?: {
    /**
     * Unit symbol
     */
    symbol: string | null;
    /**
     * Long name of the unit
     */
    name: string | null;
    /**
     * The unit dimensionality, compatible with the `pint` python package
     */
    dim: {
      [key: string]: number;
    };
  } | null;
  objectType: string;
  dynamic: boolean;
};
// ...
export type QualifiedAttribute = {
  qualifier: string | null;
  attribute: EventAttributeDefinition | ObjectAttributeDefinition;
};